Move PrefixLen and helpers to Ouroboros.Consensus.Storage.Common
So that we later can refer from
`Ouroboros.Consensus.Storage.ChainDB.Serialisation`. The dependency was
previously in the wrong direction, `Common` should not rely on `ChainDB`.
